New addition to the API, added ANDatInterface

+++ gw2DatTools/include/gw2DatTools/c_api/compression_inflateBuffer.h
@@ -19,7 +19,7 @@ extern "C" {
  *  @Return:
  *    - Pointer to the outputBuffer, NULL if it failed
  */
-GW2DATTOOLS_API uint8_t* GW2DATTOOLS_APIENTRY compression_inflateBuffer(uint32_t* iInputTab, const uint32_t iInputSize, uint32_t* ioOutputSize);
+GW2DATTOOLS_API uint8_t* GW2DATTOOLS_APIENTRY compression_inflateBuffer(uint8_t* iInputTab, const uint32_t iInputSize, uint32_t* ioOutputSize);
 
 #ifdef __cplusplus
 }
